rec-service/requirements.txt
kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
nltk==3.8.1